trading-engine/requirements.txt
# Trading Engine Dependencies - DEX Altseason Edition
# Core trading and data analysis
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0

# Cryptocurrency and DeFi
web3>=6.9.0
eth-account>=0.9.0
ccxt>=4.0.0

# API integrations
anthropic>=0.3.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Data analysis and ML
scikit-learn>=1.3.0
ta>=0.10.0
plotly>=5.15.0

# Database and caching
redis>=4.6.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

